fixed some compiler warning with float and bool

+++ src/celengine/star.cpp
@@ -986,7 +986,7 @@ float Star::getRadius() const
 
     // Use the Stefan-Boltzmann law to estimate the radius of a
     // star from surface temperature and luminosity
-    return SOLAR_RADIUS * (float) sqrt(boloLum) *
+    return (float)SOLAR_RADIUS * sqrtf(boloLum) *
         square(SOLAR_TEMPERATURE / getTemperature());
 #endif
 }
